src/main/java/me/goodjwon/springrestapistudy/accounts/AccountService.java
package me.goodjwon.springrestapistudy.accounts;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.Collection;
import java.util.Set;
import java.util.stream.Collectors;

@Service
public class AccountService implements UserDetailsService {

    final AccountRepository accountRepository;

    public AccountService(AccountRepository accountRepository) {
        this.accountRepository = accountRepository;
    }

    //우리가 사용하는 도메인을 Spring Security 가 인지 할 수 있는 형대로 바꾸어 리턴 함.
    @Override
    public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {
        Account account = accountRepository.findByUserName(userName)
                .orElseThrow(()->new UsernameNotFoundException(userName));
        return new User(account.getUserName(), account.getPassword(), authorities(account.getRoles()));
    }

    private Collection<? extends GrantedAuthority> authorities(Set<AccountRole> roles) {
        return roles.stream()
                .map(r-> new SimpleGrantedAuthority("ROLE_"+ r.name()))
                .collect(Collectors.toSet());
    }
}
